requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
